Add ValidationError to repository

+++ maker-checker/models/approval_request_repository.py
@@ -2,6 +2,10 @@ from botocore.exceptions import ClientError
 from boto3.resources.base import ServiceResource
 from boto3.dynamodb.conditions import Key, Attr  # Import Key and Attr
 from models.ApprovalRequest import ApprovalResponse, DeleteRequest
+from datetime import datetime
+
+class ValidationError(Exception):
+    pass
 
 class ApprovalRequestRepository:
     def __init__(self, db: ServiceResource) -> None:
@@ -52,21 +56,31 @@ class ApprovalRequestRepository:
             raise ValueError(e.response['Error']['Message'])
 
     def get_expired_approval_requests(self):
-        pass
-        # try:
-        #     table = self.__db.Table('approval_request')
-        #     response = table.scan(
-        #         FilterExpression=Attr("status").eq('rejected')
-        #     )
-        #     items = response.get('Items', [])
-        #     return items
-        # except ClientError as e:
-        #     raise ValueError(e.response['Error']['Message'])
+        # Check if item is not expired
+        # Untested 
+        try:
+            table = self.__db.Table('approval_request')
+            
+            # Get the current datetime
+            current_time = datetime.now()
+            
+            # Define the filter expression to check for pending items and not expired items
+            filter_expression = Attr("status").eq('pending') & Attr("request_expiry").gt(current_time)
+            
+            response = table.scan(FilterExpression=filter_expression)
+            items = response.get('Items', [])
+            return items
+        except ClientError as e:
+            raise ValueError(e.response['Error']['Message'])
 
     def create_approval_request(self, approval_request: dict):
-        table = self.__db.Table('approval_request')
-        response = table.put_item(Item=approval_request.dict())
-        return response                         #
+        try:
+            table = self.__db.Table('approval_request')
+            response = table.put_item(Item=approval_request.dict())
+            return response                         #
+        except ClientError as e:
+            raise ValueError(e.response['Error']['Message'])
+
 
     def update_approval_request(self,
                        data: dict
@@ -75,21 +89,26 @@ class ApprovalRequestRepository:
         Update function to be called only by requestor
         """
         # Retrieve the current data object from the database by UID
-        # Compare the requestor uuid with the one in the database
-        # make changes
-        table = self.__db.Table('approval_request')
-        item = table.get_item(Key={'uid': data.uid}).get('Item')
-        # Check if requestor is the same
-        assert item.get('requestor_id') == data.requestor_id, "Requestor is not the same as the original requestor!"
-        # Update the item
-        for key, value in data.dict().items():
-            # Can add additional validation before
-            if value == None:
-                pass
-            else:
-                item[key] = value
-        response = table.put_item(Item=item)
-        return response
+        try:
+            table = self.__db.Table('approval_request')
+            item = table.get_item(Key={'uid': data.uid}).get('Item')
+
+            # Compare the requestor uuid with the one in the database
+            if item.get('requestor_id') == data.approver_id:
+                raise ValidationError("Only original requestor can update request!")
+            
+            # Update the item
+            for key, value in data.dict().items():
+                # Can add additional validation before
+                if value == None:
+                    pass
+                else:
+                    item[key] = value
+            response = table.put_item(Item=item)
+            return response
+        except ClientError as e:
+            raise ValueError(e.response['Error']['Message'])
+
     
     def approve_or_reject_approval_request(self, data:ApprovalResponse):
         """
@@ -100,13 +119,16 @@ class ApprovalRequestRepository:
             item = table.get_item(Key={'uid': data.uid}).get('Item')
 
             # Check if requestor is the same as approver
-            assert item.get('requestor_id') != data.approver_id, "Requestor cannot be the same as the approver!"
-
+            if item.get('requestor_id') == data.approver_id:
+                raise ValidationError("Requestor cannot be the same as the approver!")
+            
             # Check if request has been approved / rejected / etc
-            assert item.get('status') == 'pending', "Request has already been resolved!"
+            if item.get('status') != 'pending':
+                raise ValidationError("Request has already been resolved!")
 
             # Check if item is not expired
-            assert item.get('request_expiry') > data.resolution_at, "Request has expired!"
+            if item.get('request_expiry') < data.resolution_at:
+                raise ValidationError("Request has expired!")
 
             # Update the item
             for key, value in data.model_dump().items():
@@ -115,30 +137,43 @@ class ApprovalRequestRepository:
 
             # Send out call to do whatever the request contains
             if item.get('status') == 'approved':
-                # Do something
+                # Japheth do things here
+                # Make a call to user storage to get list of all emails to send to queue
+                # Make a call to points storage to update transaction
                 pass
             response = table.put_item(Item=item)
             return response
-        except AssertionError as e:
-            print(e)
-            return e
+        except ClientError as e:
+            raise ValueError(e.response['Error']['Message'])
+
+        
         
     def withdraw_approval_request(self, data: ApprovalResponse):
+        """
+        Changes an approval object request to be withdrawn.
+        Can only be done by original requestor.
+        An expired, approved, rejected request cannot be withdrawn.
+        """
         try:
             table = self.__db.Table('approval_request')
             item = table.get_item(Key={'uid': data.uid}).get('Item')
 
             # Check if requestor is the same as approver
-            assert item.get('requestor_id') == data.approver_id, "Requestor must be the same as approver!"
+            if item.get('requestor_id') == data.approver_id:
+                raise ValidationError("Requestor cannot be the same as the approver!")
 
             # Check if request has been approved / rejected / etc
-            assert item.get('status') == 'pending', "Request has already been resolved!"
+            if item.get('status') != 'pending':
+                raise ValidationError("Request has already been resolved!")
 
             # Check if item is not expired
-            assert item.get('request_expiry') > data.resolution_at, "Request has expired!"
+            if item.get('request_expiry') < data.resolution_at:
+                raise ValidationError("Request has expired!")
 
             # Confirm its a withdraw status
-            assert data.status == 'withdrawn', "Request status must be set to withdrawn!"
+            if data.status != 'withdrawn':
+                # Do something
+                raise ValidationError("Incorrect status type for request!")
             # Update the item
             for key, value in data.model_dump().items():
                 # Can add additional validation before
@@ -147,28 +182,27 @@ class ApprovalRequestRepository:
             # Send out call to do whatever the request contains
             response = table.put_item(Item=item)
             return response
-        except AssertionError as e:
-            print(e)
-            return e
+    
+        except ClientError as e:
+            raise ValueError(e.response['Error']['Message'])
+
+
 
     def delete_approval_request(self, data:DeleteRequest):
         """
-        Not sure if we will be using this function
+        Not sure if we will be using this function or just the withdraw function
         """
         try:
             table = self.__db.Table('approval_request')
+
+            # Check if deleter is the requestor
             item = table.get_item(Key={'uid': data.uid}).get('Item')
+            # Compare the requestor uuid with the one in the database
+            if item.get('requestor_id') == data.approver_id:
+                raise ValidationError("Only original requestor can delete request!")
 
             # Send out call to do whatever the request contains
             response = table.delete_item(Key={'uid': data.uid})
             return response
-        except AssertionError as e:
-            print(e)
-            return e
-
-    # def delete_recipe(self, uid: str):
-    #     table = self.__db.Table('Recipes')      # referencing to table Recipes
-    #     response = table.delete_item(           # delete recipe using uuid
-    #         Key={'uid': uid}
-    #     )
-    #     return response
+        except ClientError as e:
+            raise ValueError(e.response['Error']['Message'])
